Navbar component with integrated search
