+++ src/style/_interactives.scss
@@ -30,7 +30,7 @@ button,
 input {
   font-family: var(--font-text);
   font-size: 1.25rem;
-  padding: 0.25rem;
+  padding: 0 0.5rem;
   margin: 0.125rem;
   border-radius: 0.5rem;
   cursor: pointer;
